requirements.txt
Django==2.2.4
django-easy-select2==1.5.6
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
